Safari-611.1.10/LayoutTests/accessibility/mac/label-element-changing-children-string-value-expected.txt

 firstfoo choice
This tests that if a label element's children change, the string value updates

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS label.role is 'AXRole: AXStaticText'
PASS initialStringValue is 'AXValue: first choice'
PASS mutatedStringValue is 'AXValue: first foo choice'
PASS successfullyParsed is true

TEST COMPLETE
